rockchip: rk3288: Enable PWM1/VDD_LOGIC at boot
While we are able to boot Endless OS on the eMMC chainloading U-Boot (in
the eMMC) from depthcharge, booting Endless OS using U-Boot (SPL and second
stage) from the SPI fails during the kernel loading.

More specifically the kernel hangs when the frequency for the PWM1 is
changed/enabled at probe time. PWM1 is used as PWM regulator (VDD) for
the logic internal to the SoC.

Looking at the configuration registers for the PWM1 in U-Boot when this
is loaded by depthcharge vs when it is loaded by the SPL it is possible
to see that PWM1 is only initialized by depthcharge whereas U-Boot
doesn't touch at all the PWM1 output that keeps being off when booting
from SPI.

Since manually enabling the PWM1 output when booting from SPI results in
a successful booting, this suggests that a way to fix the problem is to
enable the PWM1 output in the second stage U-Boot.

Even thought the PWM1 is not disabled in the DTS, its user (the
'vdd_logic' PWM regulator) is not actively used by anything, resulting
in the PWM1 not being enabled at all. Bypass this limitation simply
enabling it in the board late_init call.

Signed-off-by: Carlo Caione <carlo@endlessm.com>