rk3288: sdram: fix incorrect detection of RAM size
The 4G bitland sample is not correctly detecting the SDRAM size,
reporting 0 MB of RAM. This is due to an overflow problem where the RAM
size is saved into a 'size_t' variable.

This is still being investigated / fixed upstream (see [0]) but we can
still use a workaround to have the platform working correctly.

The workaround is based on the principle that not all the RAM is
available since part of it is reserved for MMIO. So on a 4GB platform
the actualy RAM size is (4GB-32MB), correctly fitting into the 'size_t'
variable.

[0] https://www.mail-archive.com/u-boot@lists.denx.de/msg285146.html

Signed-off-by: Carlo Caione <carlo@endlessm.com>