rk3288: Re-enable an empty banner
With commit [0] we tried to disable the U-Boot banner by disabling
CONFIG_DISPLAY_BOARDINFO_LATE. This is having the undesirable side
effect of presenting a couple of white bands when cold booting the
machine (persisting on the screen for ~1sec before disappearing when the
kernel is loaded).

This is probably due to to the fact that at boot the scanout buffer is
in some kind on uninitialized state and in this state it persists since
nothing is pushed on the screen by U-Boot.

We try to workaround this by puts-ing an empty string so that the screen
can be cleared at boot.

[0] 814506f997aba616fd8dfa33163108c530e9aadb.
    ["rk3288: Disable U-Boot banner"]

Signed-off-by: Carlo Caione <carlo@endlessm.com>